Use BigDecimal instead of Double for coin and currency values

+++ app/src/main/java/io/horizontalsystems/bankwallet/entities/LatestRate.kt
@@ -1,5 +1,6 @@
 package io.horizontalsystems.bankwallet.entities
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
-data class LatestRate(@SerializedName("rate") val value: Double, @SerializedName("date") val timestamp: Long)
+data class LatestRate(@SerializedName("rate") val value: BigDecimal, @SerializedName("date") val timestamp: Long)
